odd numbers without for loop program
