+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app', ':JavaCode'
-rootProject.name='My Application'
+rootProject.name='CodingPractice'
